requirements_test.txt
-r requirements.txt
pytest==8.3.1
pytest-cov==5.0.0
pytest-homeassistant-custom-component>=0.13
pytest-timeout==2.3.1
aiohttp-zlib-ng
requests-mock==1.12.1
